+++ branches/ng/src/MainWindow.cpp
@@ -482,8 +482,8 @@ void MainWindow::updateCurrentAccount(int id) {
 	connect(config->currentAccount(), SIGNAL(lastMessageReceived(const QString &, Account *)), this, SLOT(updateLastMessage(const QString &, Account *)));
 	disconnect(config->accounts[oldAccountId], SIGNAL(remainingRequestsUpdated(int, Account *)), 0, 0);
 	connect(config->currentAccount(), SIGNAL(remainingRequestsUpdated(int, Account *)), this, SLOT(updateRemainingRequests(int, Account *)));
-	disconnect(messageTextEdit, SIGNAL(messageEntered(const QString &, int)), 0, 0);
-	connect(messageTextEdit, SIGNAL(messageEntered(const QString &, int)), config->currentAccount(), SLOT(sendMessage(const QString &, int)));
+	disconnect(messageTextEdit, SIGNAL(messageEntered(const QString &, quint64)), 0, 0);
+	connect(messageTextEdit, SIGNAL(messageEntered(const QString &, quint64)), config->currentAccount(), SLOT(sendMessage(const QString &, quint64)));
 	messageTextEdit->setEnabled(!config->currentAccount()->sendingMessage);
 	connect(config->currentAccount(), SIGNAL(messageSent(const QString &, Account *)), this, SLOT(messageSent(const QString &, Account *)));
 	updateLastMessage(config->currentAccount()->lastMessage.text, config->currentAccount());
